requirements.txt
fastapi
uvicorn
ultralytics
torch
opencv-python-headless
pillow
python-multipart
